requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.12.0=pypi_0
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
albumentations=1.0.3=pypi_0
altair=4.1.0=pypi_0
antlr4-python3-runtime=4.8=pypi_0
anyio=3.4.0=pypi_0
apache-beam=2.30.0=pypi_0
appdirs=1.4.4=pypi_0
argon2-cffi=21.1.0=pypi_0
argparse=1.4.0=pypi_0
asgiref=3.4.1=pypi_0
astor=0.8.1=pypi_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.1=pypi_0
attrs=21.2.0=pypi_0
audioread=2.1.9=pypi_0
avro-python3=1.9.2.1=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
backports-zoneinfo=0.2.1=pypi_0
base58=2.1.1=pypi_0
blas=1.0=mkl
bleach=4.1.0=pypi_0
blinker=1.4=pypi_0
ca-certificates=2021.7.5=h06a4308_1
cachetools=4.2.1=pypi_0
certifi=2021.5.30=py38h06a4308_0
cffi=1.14.5=pypi_0
chardet=4.0.0=pypi_0
charset-normalizer=2.0.9=pypi_0
click=7.1.2=pypi_0
configparser=5.1.0=pypi_0
contextlib2=0.6.0.post1=pypi_0
crcmod=1.7=pypi_0
cryptography=3.4.7=pypi_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=pypi_0
cython=0.29.22=pypi_0
dataclasses=0.6=pypi_0
debugpy=1.4.1=py38h295c915_0
decorator=4.4.2=pypi_0
defusedxml=0.7.1=pypi_0
dill=0.3.1.1=pypi_0
distro=1.5.0=pypi_0
dm-tree=0.1.6=pypi_0
docker-pycreds=0.4.0=pypi_0
docopt=0.6.2=pypi_0
easydict=1.9=pypi_0
efficientnet-pytorch=0.7.1=pypi_0
einops=0.3.2=pypi_0
entrypoints=0.3=py38_0
fastapi=0.70.0=pypi_0
fastavro=1.4.1=pypi_0
filelock=3.0.12=pypi_0
flask=2.0.1=pypi_0
flatbuffers=1.12=pypi_0
freetype=2.10.4=h5ab3b9f_0
frozenlist=1.2.0=pypi_0
fsspec=2021.11.1=pypi_0
future=0.18.2=pypi_0
gast=0.4.0=pypi_0
gdown=3.13.0=pypi_0
gin-config=0.4.0=pypi_0
gitdb=4.0.9=pypi_0
gitpython=3.1.24=pypi_0
google-api-core=1.30.0=pypi_0
google-api-python-client=2.10.0=pypi_0
google-auth=1.28.0=pypi_0
google-auth-httplib2=0.1.0=pypi_0
google-auth-oauthlib=0.4.3=pypi_0
google-cloud-bigquery=2.20.0=pypi_0
google-cloud-core=1.7.0=pypi_0
google-crc32c=1.1.2=pypi_0
google-pasta=0.2.0=pypi_0
google-resumable-media=1.3.1=pypi_0
googleapis-common-protos=1.53.0=pypi_0
grpcio=1.34.1=pypi_0
h11=0.12.0=pypi_0
h5py=3.1.0=pypi_0
hdfs=2.6.0=pypi_0
httplib2=0.19.1=pypi_0
huggingface-hub=0.2.0=pypi_0
idna=2.10=pypi_0
imageio=2.9.0=pypi_0
imageio-ffmpeg=0.4.4=pypi_0
imgaug=0.4.0=pypi_0
importlib-metadata=4.5.0=pypi_0
importlib-resources=5.1.4=pypi_0
intel-openmp=2020.2=254
ipykernel=6.2.0=py38h06a4308_1
ipython=7.27.0=py38hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pypi_0
itsdangerous=2.0.1=pypi_0
jedi=0.18.0=py38h06a4308_1
jeepney=0.6.0=pypi_0
jinja2=3.0.1=pypi_0
joblib=1.0.1=pypi_0
jpeg=9b=h024ee3a_2
jsonschema=4.2.1=pypi_0
jupyter_client=7.0.1=pyhd3eb1b0_0
jupyter_core=4.7.1=py38h06a4308_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.2=pypi_0
kaggle=1.5.12=pypi_0
keras=2.4.3=pypi_0
keras-nightly=2.5.0.dev2021032900=pypi_0
keras-preprocessing=1.1.2=pypi_0
keyring=23.0.1=pypi_0
kiwisolver=1.3.1=pypi_0
launchpadlib=1.10.13=pypi_0
lazr-restfulclient=0.14.3=pypi_0
lazr-uri=1.0.5=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
librosa=0.8.0=pypi_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuv=1.40.0=h7b6447c_0
llvmlite=0.36.0=pypi_0
lvis=0.5.3=pypi_0
lxml=4.6.3=pypi_0
lz4-c=1.9.3=h2531618_0
markdown=3.3.4=pypi_0
markupsafe=2.0.1=pypi_0
matplotlib=3.3.4=pypi_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.3.0=py38h54f3939_0
mkl_random=1.1.1=py38h0573a6f_0
mlconfig=0.1.0=pypi_0
moviepy=1.0.3=pypi_0
mtcnn=0.1.0=pypi_0
multidict=5.2.0=pypi_0
munch=2.5.0=pypi_0
nbclient=0.5.9=pypi_0
nbconvert=6.3.0=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.6.3=pypi_0
ninja=1.10.2=py38hff7bd54_0
nltk=3.6.5=pypi_0
notebook=6.4.6=pypi_0
numba=0.53.0=pypi_0
numpy=1.19.5=pypi_0
numpy-base=1.19.2=py38hfa32c7d_0
oauth2client=4.1.3=pypi_0
oauthlib=3.1.0=pypi_0
object-detection=0.1=pypi_0
olefile=0.46=py_0
omegaconf=2.1.1=pypi_0
opencv-contrib-python=4.5.1.48=pypi_0
opencv-python=4.5.1.48=pypi_0
opencv-python-headless=4.5.2.54=pypi_0
openssl=1.1.1l=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=20.9=pypi_0
pandas=1.2.3=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.2=pyhd3eb1b0_0
pathtools=0.1.2=pypi_0
pbr=5.6.0=pypi_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.1.2=pypi_0
pip=21.0.1=py38h06a4308_0
pipdeptree=2.2.0=pypi_0
plotly=4.14.3=pypi_0
pooch=1.3.0=pypi_0
portalocker=2.0.0=pypi_0
pretrainedmodels=0.7.4=pypi_0
proglog=0.1.9=pypi_0
prometheus-client=0.12.0=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.17=pyhca03da5_0
proto-plus=1.18.1=pypi_0
protobuf=3.15.6=pypi_0
psutil=5.8.0=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
py-cpuinfo=8.0.0=pypi_0
pyarrow=3.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.2=pypi_0
pycparser=2.20=pypi_0
pydantic=1.8.2=pypi_0
pydeck=0.7.1=pypi_0
pydeprecate=0.3.1=pypi_0
pydot=1.4.2=pypi_0
pygments=2.10.0=pyhd3eb1b0_0
pymongo=3.11.4=pypi_0
pympler=0.9=pypi_0
pyparsing=2.4.7=pypi_0
pyrsistent=0.18.0=pypi_0
pysocks=1.7.1=pypi_0
python=3.8.8=hdb3f193_4
python-dateutil=2.8.1=pypi_0
python-levenshtein=0.12.2=pypi_0
python-multipart=0.0.5=pypi_0
python-slugify=5.0.2=pypi_0
pytorch=1.7.1=py3.8_cuda10.1.243_cudnn7.6.3_0
pytorch-lightning=1.5.5=pypi_0
pytorch-model-summary=0.1.2=pypi_0
pytz=2021.1=pypi_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pywavelets=1.1.1=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=22.2.1=py38h295c915_1
readline=8.1=h27cfd23_0
regex=2021.11.10=pypi_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
resampy=0.2.2=pypi_0
retrying=1.3.3=pypi_0
rsa=4.7.2=pypi_0
sacrebleu=1.5.1=pypi_0
sacremoses=0.0.46=pypi_0
scikit-image=0.18.3=pypi_0
scikit-learn=0.24.1=pypi_0
scipy=1.6.1=pypi_0
seaborn=0.11.2=pypi_0
secretstorage=3.3.1=pypi_0
semantic-version=2.8.5=pypi_0
send2trash=1.8.0=pypi_0
sentencepiece=0.1.96=pypi_0
sentry-sdk=1.4.3=pypi_0
seqeval=1.2.2=pypi_0
setuptools=52.0.0=py38h06a4308_0
shapely=1.8.0=pypi_0
shortuuid=1.0.8=pypi_0
six=1.15.0=py38h06a4308_0
smmap=5.0.0=pypi_0
sniffio=1.2.0=pypi_0
soundfile=0.10.3.post1=pypi_0
sqlite=3.35.1=hdfb4753_0
starlette=0.16.0=pypi_0
streamlit=1.2.0=pypi_0
subprocess32=3.5.4=pypi_0
tensorboard=2.5.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorflow=2.5.0=pypi_0
tensorflow-addons=0.13.0=pypi_0
tensorflow-datasets=4.3.0=pypi_0
tensorflow-estimator=2.5.0=pypi_0
tensorflow-hub=0.12.0=pypi_0
tensorflow-metadata=1.1.0=pypi_0
tensorflow-model-optimization=0.6.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.12.1=pypi_0
testpath=0.5.0=pypi_0
testresources=2.0.1=pypi_0
text-unidecode=1.3=pypi_0
tf-models-official=2.5.0=pypi_0
tf-slim=1.1.0=pypi_0
threadpoolctl=2.1.0=pypi_0
tifffile=2021.8.30=pypi_0
tk=8.6.10=hbc83047_0
tokenizers=0.10.3=pypi_0
toml=0.10.2=pypi_0
toolz=0.11.2=pypi_0
torch=1.7.0+cu101=pypi_0
torch-tb-profiler=0.2.1=pypi_0
torchaudio=0.7.0=pypi_0
torchmetrics=0.6.1=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.9.0+cu111=pypi_0
tornado=6.1=py38h27cfd23_0
tqdm=4.59.0=pypi_0
traitlets=5.0.5=pyhd3eb1b0_0
transformers=4.12.5=pypi_0
typeguard=2.12.1=pypi_0
typing_extensions=3.7.4.3=pyha847dfd_0
tzdata=2021.5=pypi_0
tzlocal=4.1=pypi_0
uritemplate=3.0.1=pypi_0
urllib3=1.26.4=pypi_0
uvicorn=0.16.0=pypi_0
validators=0.18.2=pypi_0
wadllib=1.3.5=pypi_0
wandb=0.12.6=pypi_0
watchdog=2.1.6=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=pypi_0
werkzeug=2.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.2=pypi_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h7b6447c_0
yarl=1.7.2=pypi_0
yaspin=2.1.0=pypi_0
zeromq=4.3.4=h2531618_0
zipp=3.4.1=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
